requirements.txt
playwright>=1.44.0
pydantic>=2.0
python-docx
pdfplumber
pyyaml
